Narrowed down the problem
I got to thinking maybe the issue isn't with .NET. I wanted to narrow it
down, so I just ran the eval by itself, (no JavaScript) and it pulls the
description field just fine when it contains a carriage return. It
doesn't display the carriage return, but it works.  As soon as I wrap it
in that little script to remove it if it's blank, it won't display
anything if the field value contains a carriage return.